KNIME-1309: Support all port objects, not only data tables (#133)
* KNIME-1309 Added: support of dynamic ports and arbitrary port types
* Dependency fix for RDKit Nodes Wizard
---------
