+++ org.rdkit.knime.wizards/META-INF/MANIFEST.MF
@@ -17,7 +17,7 @@ Require-Bundle: org.eclipse.core.runtime;bundle-version="[3.20.0,4.0.0)",
  org.eclipse.ui.views.properties.tabbed;bundle-version="[3.9.0,4.0.0)",
  org.eclipse.gef;bundle-version="[3.11.0,4.0.0)",
  org.eclipse.pde;bundle-version="[3.13.0,4.0.0)",
- org.apache.commons.io;bundle-version="[2.8.0,3.0.0)"
+ org.apache.commons.io;bundle-version="[2.6.0,3.0.0)"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
 Bundle-ClassPath: .
